refactor(view): modularize inline views
