+++ tests/module_loading_spec.lua
@@ -118,7 +118,7 @@ describe("Explorer submodules", function()
   end)
 end)
 
--- ── side-by-side view ──────────────────────────────────────────────────
+-- ── view engines ───────────────────────────────────────────────────────
 describe("Side-by-side view module", function()
   it("loads the facade and exports its existing API", function()
     local ok, mod = pcall(require, "codediff.ui.view.side_by_side")
@@ -129,6 +129,18 @@ describe("Side-by-side view module", function()
     assert.is_function(mod.show_deleted_file)
     assert.is_function(mod.show_added_virtual_file)
     assert.is_function(mod.show_deleted_virtual_file)
+    assert.is_function(mod.show_welcome)
+  end)
+end)
+
+describe("Inline view module", function()
+  it("loads the facade and exports its existing API", function()
+    local ok, mod = pcall(require, "codediff.ui.view.inline_view")
+    assert.is_true(ok, "Failed to require codediff.ui.view.inline_view: " .. tostring(mod))
+    assert.is_function(mod.create)
+    assert.is_function(mod.update)
+    assert.is_function(mod.rerender)
+    assert.is_function(mod.show_single_file)
     assert.is_function(mod.show_welcome)
   end)
 end)
